FileManager tests

+++ Sriracha.Deploy.Data/Build/BuildImpl/FileManager.cs
@@ -28,39 +28,90 @@ namespace Sriracha.Deploy.Data.Build.BuildImpl
 
 		public DeployFile CreateFile(string fileName, byte[] fileData)
 		{
+            if(string.IsNullOrEmpty(fileName))
+            {
+                throw new ArgumentNullException("Missing fileName");
+            }
+            if(fileData == null)
+            {
+                throw new ArgumentNullException("Missing fileData");
+            }
+            if(fileData.Length == 0)
+            {
+                throw new ArgumentException("fileData is empty");
+            }
 			return _fileRepository.CreateFile(fileName, fileData);
 		}
 
 		public DeployFile UpdateFile(string fileId, string fileName, byte[] fileData)
 		{
+            if(string.IsNullOrEmpty(fileId))
+            {
+                throw new ArgumentNullException("Missing fileId");
+            }
+            if(string.IsNullOrEmpty(fileName))
+            {
+                throw new ArgumentNullException("Missing fileName");
+            }
+            if(fileData == null)
+            {
+                throw new ArgumentNullException("Missing fileData");
+            }
+            if(fileData.Length == 0)
+            {
+                throw new ArgumentException("fileData is empty");
+            }
 			return _fileRepository.UpdateFile(fileId, fileName, fileData);
 		}
 
 		public DeployFile GetFile(string fileId)
 		{
+            if(string.IsNullOrEmpty(fileId))
+            {
+                throw new ArgumentNullException("Missing fileId");
+            }
 			return _fileRepository.GetFile(fileId);
 		}
 
         public byte[] GetFileData(string fileId)
         {
+            if (string.IsNullOrEmpty(fileId))
+            {
+                throw new ArgumentNullException("Missing fileId");
+            }
             return _fileRepository.GetFileData(fileId);
         }
 
         public Stream GetFileDataStream(string fileId)
         {
+            if (string.IsNullOrEmpty(fileId))
+            {
+                throw new ArgumentNullException("Missing fileId");
+            }
             return _fileRepository.GetFileDataStream(fileId);
         }
 
 		public void DeleteFile(string fileId)
 		{
-			_fileRepository.DeleteFile(fileId);
+            if (string.IsNullOrEmpty(fileId))
+            {
+                throw new ArgumentNullException("Missing fileId");
+            }
+            _fileRepository.DeleteFile(fileId);
 		}
 
 		public void ExportFile(string fileId, string targetFilePath)
 		{
-			var file = this.GetFile(fileId);
-			var data = _fileRepository.GetFileData(fileId);
-			_fileWriter.WriteBytes(targetFilePath, data);
+            if (string.IsNullOrEmpty(fileId))
+            {
+                throw new ArgumentNullException("Missing fileId");
+            }
+            var data = _fileRepository.GetFileData(fileId);
+            if (data == null)
+            {
+                throw new ArgumentException("File not found for ID " + fileId);
+            }
+            _fileWriter.WriteBytes(targetFilePath, data);
 		}
     }
 }
